+++ src/reducer.js
@@ -1,5 +1,4 @@
 import compose from 'fkit/dist/compose'
-import curry from 'fkit/dist/curry'
 import id from 'fkit/dist/id'
 import inc from 'fkit/dist/inc'
 import last from 'fkit/dist/last'
@@ -44,6 +43,19 @@ function trackImpressions (promos, ts, user) {
 }
 
 /**
+ * Tracks the engagements for the given promo.
+ *
+ * @private
+ */
+function trackEngagements (promo, ts, user) {
+  return compose(
+    updateEntity(['engagements', 'campaigns', promo.campaignId], ts),
+    updateEntity(['engagements', 'groups', promo.groupId], ts),
+    updateEntity(['engagements', 'promos', promo.promoId], ts)
+  )(user)
+}
+
+/**
  * Increments the counter and sets the timestamp for the given entity.
  *
  * @private
@@ -65,19 +77,22 @@ function updateEntity (keyPath, ts) {
 /**
  * Applies an event to the current state to yield a new state.
  *
- * @param {Object} context The context object.
+ * @param {Array} promos The list of promos.
+ * @param {Window} window The window object.
  * @param {Object} state The current state.
  * @param {Object} event The event.
  * @returns A new state.
  */
-function reducer (context, state, event) {
+function reducer (promos, window, state, event) {
   const ts = timestamp()
-  let { window, promos } = context
   let { user } = state
 
   if (event.type === 'visit') {
     // Increment the number of user visits.
     user = incrementVisits(user)
+  } else if (event.type === 'click') {
+    // Add promo to the list of engagements.
+    user = trackEngagements(event.promo, ts, user)
   } else if (event.type === 'close') {
     // Add promo to the list of blocked promos.
     user = blockPromo(event.promo, ts, user)
@@ -92,4 +107,4 @@ function reducer (context, state, event) {
   return { promos: placedPromos, user }
 }
 
-export default curry(reducer)
+export default reducer